fix: fix and fmt workflows, update readme to new convention
Run Automatic PRs on `dev` pushes instead of what we had before, and add
commit history as part of the PR body. And just added a few spaces to
make it more readable in all of the workflows.

Changed contribution guide to conform to us just using a `dev` branch
rather than what we had before.