feat: Add more documentation, add install script, add more docker, do some small fixes
Added installation instructions, aswell as extend building instructions
through choices of using Just or not.

Added an install script you should be able to invoke with curl or wget,
that will let you quickly download everything that's needed to run
PetRing in a docker container.

Added a `docker-compose.override.yml` specifically to make it easier on
people who use traefik.

Restored the migration justfile to a working state