fix: change url, make `download.sh` actually work, fix migration justfile
The url in `README.md` was wrong, as it pointed to a nonexistend
`install.sh`, so fixed that.

Did some changes to `download.sh` that makes it actually copy over the
proper files, and see if the person has just installed or not.

Removed "api/" in the migration recipe in the root `justfile` because
`migration` is no longer a subdirectory of `api/`